Issue #633: Moves class FXWrapper to package com.canoo.platform.remoting.client.javafx Also fixes import statement in platform-examples
